separate functions shared with ironpython into their own file
