+++ wordpool/__init__.py
@@ -1,18 +1,8 @@
-try:
-    import numpy as np
-except ImportError:
-    np = None
+import numpy as np
+import pandas as pd
+from .nopandas import assign_list_numbers_from_word_list
+from pkg_resources import resource_filename, resource_listdir
 
-try:
-    import pandas as pd
-except ImportError:
-    pd = None
-
-try:
-    from pkg_resources import resource_filename, resource_listdir
-except ImportError:
-    resource_filename = None
-    resource_listdir = None
 
 __version__ = "0.4.1"
 
@@ -39,22 +29,8 @@ def load(filename, from_data_package=True):
         src = filename
     return pd.read_table(src)
 
-def assign_list_numbers(words, number_of_lists, start=0):
-    """Assign or reassign list numbers to all words in the pool.
-    
-    :param words: Input word pool- Either a dataframe or a list of {'word': "word"} dictionaries.
-    :param int number_of_lists: Total number of lists.
-    :param int start: Start number for lists.
-    :returns: Word pool with list numbers assigned.  This is a list of dictionaries if this was called with a list of dictionaries, otherwise it is a dataframe.
-        
-    """
-    if (type(words) == list):
-        return _assign_list_numbers_no_pandas(words, number_of_lists, start=start)
-    else:
-        return _assign_list_numbers_pandas(words, number_of_lists, start=start)
 
-
-def _assign_list_numbers_pandas(df, n_lists, start=0):
+def assign_list_numbers(df, n_lists, start=0):
     """Assign or reassign list numbers to all words in the pool.
 
     :param pd.DataFrame df: Input word pool
@@ -66,26 +42,10 @@ def _assign_list_numbers_pandas(df, n_lists, start=0):
     assert len(df) % n_lists == 0
     
     pool_list = pool_dataframe_to_pool_list(df)
-    pool_list = _assign_list_numbers_no_pandas(pool_list, n_lists, start=start)
+    pool_list = assign_list_numbers_from_word_list(pool_list, n_lists, start=start)
     df = pool_list_to_pool_dataframe(pool_list)
 
     return df
-
-def _assign_list_numbers_no_pandas(all_words, number_of_lists, start=0):
-    """takes a list of dictionaries with just words and adds listnos.
-        
-        :param all_words: a list of dictionaries of all the words to assign numbers to
-        :param number_of_lists: how many lists should the words be divided into
-        :returns a list of (word, list_number) pairs.
-        
-        """
-    if ((len(all_words))%number_of_lists != 0):
-        raise ValueError("The number of words must be evenly divisible by the number of lists.")
-    
-    length_of_each_list = len(all_words)/number_of_lists
-    for i in range(len(all_words)):
-        all_words[i]['listno'] = (i/length_of_each_list) + start
-    return all_words
 
 
 def pool_dataframe_to_pool_list(pool_dataframe):
